requirements.txt
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
langchain-core==0.1.23
langchain-text-splitters==0.3.8
transformers==4.33.3
torch==2.7.1
torchvision== 0.22.1
torchaudio==2.2.0
faiss-cpu==1.7.4
llama-index-core==0.12.52
llama-index-embeddings-huggingface==0.5.5
llama-index-instrumentation==0.3.0
llama-index-llms-ollama==0.6.2
llama-index-workflows==1.1.0
bangla-pdf-ocr==0.1.1
dataclasses-json==0.6.7
langdetect==1.0.9
regex==2024.11.6
nltk==3.9.1
ftfy==6.1.3
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.5.0
requests==2.32.4
ollama==0.1.7
accelerate==0.21.1
scikit-learn==1.6.1
tqdm==4.67.1
loguru==0.7.0
python-dotenv==1.0.0
jupyter==1.0.0
ipykernel==6.26.0 
onnx==1.17.0
huggingface_hub==0.21.4
sentence-transformers==2.2.2 
numpy==1.23.5